requirements.txt
Flask
Gunicorn
requests
pyyaml
python-dotenv